CLA0003-970 - Data Import On Demand
Well, functionally got it right first time, but didnt account for the files that are imported are within an app named api and then they're processed by an app named worker, so the fil path doesn't match.

Solved this by moving the EFS storage out to app level and named it storage, on both api and worker.